fix: move reusable function into utiil
