+++ addons/addon-base-raas/packages/base-raas-services/lib/data-source/access-strategy/legacy/environment-resource-service.js
@@ -17,6 +17,17 @@ const _ = require('lodash');
 const Service = require('@aws-ee/base-services-container/lib/service');
 const { processInBatches } = require('@aws-ee/base-services/lib/helpers/utils');
 
+const {
+  updateS3BucketPolicy,
+  getStatementParamsFn,
+  listStatementParamsFn,
+  putStatementParamsFn,
+  addEmptyPrincipalIfNotPresent,
+  getRootArnForAccount,
+  createAllowStatement,
+  addAccountToStatement,
+} = require('../../../helpers/utils');
+
 const { parseS3Arn } = require('../../../helpers/s3-arn');
 const { isOpenData } = require('../../../study/helpers/entities/study-methods');
 
@@ -24,41 +35,6 @@ const settingKeys = {
   studyDataBucketName: 'studyDataBucketName',
   studyDataKmsKeyArn: 'studyDataKmsKeyArn',
   studyDataKmsPolicyWorkspaceSid: 'studyDataKmsPolicyWorkspaceSid',
-};
-
-const listStatementParamsFn = (bucket, prefix) => {
-  return {
-    statementId: `List:${prefix}`,
-    resource: `arn:aws:s3:::${bucket}`,
-    actions: ['s3:ListBucket'],
-    condition: {
-      StringLike: {
-        's3:prefix': [`${prefix}*`],
-      },
-    },
-  };
-};
-
-const getStatementParamsFn = (bucket, prefix) => {
-  return {
-    statementId: `Get:${prefix}`,
-    resource: [`arn:aws:s3:::${bucket}/${prefix}*`],
-    actions: ['s3:GetObject'],
-  };
-};
-
-const putStatementParamsFn = (bucket, prefix) => {
-  return {
-    statementId: `Put:${prefix}`,
-    resource: [`arn:aws:s3:::${bucket}/${prefix}*`],
-    actions: [
-      's3:AbortMultipartUpload',
-      's3:ListMultipartUploadParts',
-      's3:PutObject',
-      's3:PutObjectAcl',
-      's3:DeleteObject',
-    ],
-  };
 };
 
 /**
@@ -285,12 +261,14 @@ class EnvironmentResourceService extends Service {
           study,
           s3BucketName,
           statementParamFunctions,
-          oldStatement => this.addAccountToStatement(oldStatement, memberAccountId),
+          oldStatement => addAccountToStatement(oldStatement, memberAccountId),
         );
         return revisedStatementsPerStudy;
       }),
     );
-    await this.updateS3BucketPolicy(s3BucketName, s3Policy, revisedStatements);
+
+    const s3Client = await this.getS3();
+    await updateS3BucketPolicy(s3Client, s3BucketName, s3Policy, revisedStatements);
 
     // Write audit event
     await this.audit(requestContext, { action: 'add-to-bucket-policy', body: s3Policy });
@@ -318,7 +296,9 @@ class EnvironmentResourceService extends Service {
         return revisedStatementsPerStudy;
       }),
     );
-    await this.updateS3BucketPolicy(s3BucketName, s3Policy, revisedStatements);
+
+    const s3Client = await this.getS3();
+    await updateS3BucketPolicy(s3Client, s3BucketName, s3Policy, revisedStatements);
 
     // Write audit event
     await this.audit(requestContext, { action: 'remove-from-bucket-policy', body: s3Policy });
@@ -326,9 +306,7 @@ class EnvironmentResourceService extends Service {
 
   // @private
   async addToKmsKeyPolicy(requestContext, memberAccountId) {
-    await this.updateKMSPolicy(environmentStatement =>
-      this.addAccountToStatement(environmentStatement, memberAccountId),
-    );
+    await this.updateKMSPolicy(environmentStatement => addAccountToStatement(environmentStatement, memberAccountId));
 
     // Write audit event
     await this.audit(requestContext, { action: 'add-to-KmsKey-policy', body: memberAccountId });
@@ -398,43 +376,12 @@ class EnvironmentResourceService extends Service {
   }
 
   // @private
-  addAccountToStatement(oldStatement, memberAccountId) {
-    const principal = this.getRootArnForAccount(memberAccountId);
-    const statement = this.addEmptyPrincipalIfNotPresent(oldStatement);
-    if (Array.isArray(statement.Principal.AWS)) {
-      // add the principal if it doesn't exist already
-      if (!statement.Principal.AWS.includes(principal)) {
-        statement.Principal.AWS.push(principal);
-      }
-    } else if (statement.Principal.AWS !== principal) {
-      statement.Principal.AWS = [statement.Principal.AWS, principal];
-    }
-    return statement;
-  }
-
-  // @private
   removeAccountFromStatement(oldStatement, memberAccountId) {
-    const principal = this.getRootArnForAccount(memberAccountId);
-    const statement = this.addEmptyPrincipalIfNotPresent(oldStatement);
+    const principal = getRootArnForAccount(memberAccountId);
+    const statement = addEmptyPrincipalIfNotPresent(oldStatement);
     if (Array.isArray(statement.Principal.AWS)) {
       statement.Principal.AWS = statement.Principal.AWS.filter(oldPrincipal => oldPrincipal !== principal);
     } else if (statement.Principal.AWS === principal) {
-      statement.Principal.AWS = [];
-    }
-    return statement;
-  }
-
-  // @private
-  getRootArnForAccount(memberAccountId) {
-    return `arn:aws:iam::${memberAccountId}:root`;
-  }
-
-  // @private
-  addEmptyPrincipalIfNotPresent(statement) {
-    if (!statement.Principal) {
-      statement.Principal = {};
-    }
-    if (!statement.Principal.AWS) {
       statement.Principal.AWS = [];
     }
     return statement;
@@ -446,7 +393,7 @@ class EnvironmentResourceService extends Service {
       const statementParams = statementParameterFn(bucket, study.prefix);
       let oldStatement = s3Policy.Statement.find(statement => statement.Sid === statementParams.statementId);
       if (!oldStatement) {
-        oldStatement = this.createAllowStatement(
+        oldStatement = createAllowStatement(
           statementParams.statementId,
           statementParams.actions,
           statementParams.resource,
@@ -457,40 +404,6 @@ class EnvironmentResourceService extends Service {
       return newStatement;
     });
     return revisedStatementsPerStudy;
-  }
-
-  // @private
-  createAllowStatement(statementId, actions, resource, condition) {
-    const baseAllowStatement = {
-      Sid: statementId,
-      Effect: 'Allow',
-      Principal: { AWS: [] },
-      Action: actions,
-      Resource: resource,
-    };
-    if (condition) {
-      baseAllowStatement.Condition = condition;
-    }
-    return baseAllowStatement;
-  }
-
-  // @private
-  async updateS3BucketPolicy(s3BucketName, s3Policy, revisedStatements) {
-    const s3Client = await this.getS3();
-
-    // remove all the old statements from s3Policy that have changed
-    const revisedStatementIds = revisedStatements.flat().map(statement => statement.Sid);
-    s3Policy.Statement = s3Policy.Statement.filter(statement => !revisedStatementIds.includes(statement.Sid));
-
-    // add all the revised statements to the s3Policy
-    revisedStatements.flat().forEach(statement => {
-      // Only add updated statement if it contains principals (otherwise leave it out)
-      if (statement.Principal.AWS.length > 0) {
-        s3Policy.Statement.push(statement);
-      }
-    });
-    // Update S3 bucket policy
-    await s3Client.putBucketPolicy({ Bucket: s3BucketName, Policy: JSON.stringify(s3Policy) }).promise();
   }
 
   // @private
