+++ addons/addon-base-raas/packages/base-raas-services/lib/helpers/utils.js
@@ -39,9 +39,110 @@ function normalizeStudyFolder(str = '') {
   return _.endsWith(resolved, '/') ? resolved : `${resolved}/`;
 }
 
+async function updateS3BucketPolicy(s3Client, s3BucketName, s3Policy, revisedStatements) {
+  // remove all the old statements from s3Policy that have changed
+  const revisedStatementIds = revisedStatements.flat().map(statement => statement.Sid);
+  s3Policy.Statement = s3Policy.Statement.filter(statement => !revisedStatementIds.includes(statement.Sid));
+
+  // add all the revised statements to the s3Policy
+  revisedStatements.flat().forEach(statement => {
+    // Only add updated statement if it contains principals (otherwise leave it out)
+    if (statement.Principal.AWS.length > 0) {
+      s3Policy.Statement.push(statement);
+    }
+  });
+  // Update S3 bucket policy
+  await s3Client.putBucketPolicy({ Bucket: s3BucketName, Policy: JSON.stringify(s3Policy) }).promise();
+}
+
+function createAllowStatement(statementId, actions, resource, condition) {
+  const baseAllowStatement = {
+    Sid: statementId,
+    Effect: 'Allow',
+    Principal: { AWS: [] },
+    Action: actions,
+    Resource: resource,
+  };
+  if (condition) {
+    baseAllowStatement.Condition = condition;
+  }
+  return baseAllowStatement;
+}
+
+function getRootArnForAccount(memberAccountId) {
+  return `arn:aws:iam::${memberAccountId}:root`;
+}
+
+function addEmptyPrincipalIfNotPresent(statement) {
+  if (!statement.Principal) {
+    statement.Principal = {};
+  }
+  if (!statement.Principal.AWS) {
+    statement.Principal.AWS = [];
+  }
+  return statement;
+}
+
+const getStatementParamsFn = (bucket, prefix) => {
+  return {
+    statementId: `Get:${prefix}`,
+    resource: [`arn:aws:s3:::${bucket}/${prefix}*`],
+    actions: ['s3:GetObject'],
+  };
+};
+
+const listStatementParamsFn = (bucket, prefix) => {
+  return {
+    statementId: `List:${prefix}`,
+    resource: `arn:aws:s3:::${bucket}`,
+    actions: ['s3:ListBucket'],
+    condition: {
+      StringLike: {
+        's3:prefix': [`${prefix}*`],
+      },
+    },
+  };
+};
+
+const putStatementParamsFn = (bucket, prefix) => {
+  return {
+    statementId: `Put:${prefix}`,
+    resource: [`arn:aws:s3:::${bucket}/${prefix}*`],
+    actions: [
+      's3:AbortMultipartUpload',
+      's3:ListMultipartUploadParts',
+      's3:PutObject',
+      's3:PutObjectAcl',
+      's3:DeleteObject',
+    ],
+  };
+};
+
+function addAccountToStatement(oldStatement, memberAccountId) {
+  const principal = getRootArnForAccount(memberAccountId);
+  const statement = addEmptyPrincipalIfNotPresent(oldStatement);
+  if (Array.isArray(statement.Principal.AWS)) {
+    // add the principal if it doesn't exist already
+    if (!statement.Principal.AWS.includes(principal)) {
+      statement.Principal.AWS.push(principal);
+    }
+  } else if (statement.Principal.AWS !== principal) {
+    statement.Principal.AWS = [statement.Principal.AWS, principal];
+  }
+  return statement;
+}
+
 module.exports = {
   generateId,
   chopRight,
   chopLeft,
   normalizeStudyFolder,
+  updateS3BucketPolicy,
+  createAllowStatement,
+  getRootArnForAccount,
+  addEmptyPrincipalIfNotPresent,
+  getStatementParamsFn,
+  listStatementParamsFn,
+  putStatementParamsFn,
+  addAccountToStatement,
 };
